Function to replace relative resources by epub://
Resources like styles or images are linked inside the xml doc with a
relative path using as base the doc path.

To render these resources in widgets like webkit this new function replaces
with a the scheme, epub://, and converts these paths from relative to
absolute so we can find these resources easily with the get_resource
function.