Added GepubWidget class to show the content with webkit
This new widget shows the epub content using webkit2gtk. The new widget
inherites from GtkBox and at first it only shows a WebKitWebView with the
current doc page, but the idea is to be able to paginate and show two pages
at the same time, for the landscape view.